Revert "Rename app to main for railway deployment"
